requirements.txt
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.95.2
h11==0.14.0
httptools==0.6.4
idna==3.10
numpy==2.0.2
pandas==2.2.3
pydantic==1.10.21
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.31.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tabula-py==2.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1